requirements.txt
flask
flask_cors
flask_session
pillow
ffmpeg-python
pydub
pandoc
chardet
unoserver
py7zr
rarfile